+++ R/RcppExports.R
@@ -183,11 +183,54 @@ lin_costs_logistic <- function(x, y, w, min_seg, indices) {
     .Call('_RcppDynProg_lin_costs_logistic', PACKAGE = 'RcppDynProg', x, y, w, min_seg, indices)
 }
 
-#' solve_interval_partition interval partition problem.
+#' solve_interval_partition interval partition problem with a bound on number of steps.
+#' 
+#' Solve a for a minimal cost partition of the integers [1,...,nrow(x)] problem where for j>=i x(i,j).
+#' is the cost of choosing the partition element [i,...,j]. 
+#' Returned solution is an ordered vector v of length k<=kmax where: v[1]==1, v[k]==nrow(x)+1, and the 
+#' partition is of the form [v[i], v[i+1]) (intervals open on the right).
+#' 
+#' @param x NumericMatix, for j>=i x(i,j) is the cost of partition element [i,...,j] (inclusive).
+#' @param kmax int, maximum number of segments in solution. 
+#' @return dynamic program solution.
+#' 
+#' @examples
+#' 
+#' costs <- matrix(c(1.5, NA ,NA ,1 ,0 , NA, 5, -1, 1), nrow = 3)
+#' solve_interval_partition(costs, nrow(costs))
+#'
+#' @export
+solve_interval_partition_k <- function(x, kmax) {
+    .Call('_RcppDynProg_solve_interval_partition_k', PACKAGE = 'RcppDynProg', x, kmax)
+}
+
+#' solve_interval_partition interval partition problem, no boun on the number of steps.
+#' 
+#' Not working yet.
 #' 
 #' Solve a for a minimal cost partition of the integers [1,...,nrow(x)] problem where for j>=i x(i,j).
 #' is the cost of choosing the partition element [i,...,j]. 
 #' Returned solution is an ordered vector v of length k where: v[1]==1, v[k]==nrow(x)+1, and the 
+#' partition is of the form [v[i], v[i+1]) (intervals open on the right).
+#' 
+#' @param x NumericMatix, for j>=i x(i,j) is the cost of partition element [i,...,j] (inclusive).
+#' @return dynamic program solution.
+#' 
+#' @examples
+#' 
+#' costs <- matrix(c(1.5, NA ,NA ,1 ,0 , NA, 5, -1, 1), nrow = 3)
+#' solve_interval_partition(costs, nrow(costs))
+#'
+#' @export
+solve_interval_partition_no_k <- function(x) {
+    .Call('_RcppDynProg_solve_interval_partition_no_k', PACKAGE = 'RcppDynProg', x)
+}
+
+#' solve_interval_partition interval partition problem.
+#' 
+#' Solve a for a minimal cost partition of the integers [1,...,nrow(x)] problem where for j>=i x(i,j).
+#' is the cost of choosing the partition element [i,...,j]. 
+#' Returned solution is an ordered vector v of length k<=kmax where: v[1]==1, v[k]==nrow(x)+1, and the 
 #' partition is of the form [v[i], v[i+1]) (intervals open on the right).
 #' 
 #' @param x NumericMatix, for j>=i x(i,j) is the cost of partition element [i,...,j] (inclusive).