+++ NEWS.md
@@ -1,7 +1,8 @@
 
-# RcppDynProg 0.1.1 2019/01/19
+# RcppDynProg 0.1.1 2019/01/21
 
  * Add logistic scoring.
+ * No-k bound variation.
 
 # RcppDynProg 0.1.0 2019/01/01